Rename library modules to the endpoints they represent.
