+++ src/variables.ts
